+++ spec/integration_helper.rb
@@ -1,4 +1,4 @@
-require 'telegram/bot/rspec/integration'
+require 'telegram/bot/rspec/integration/rails'
 require 'action_controller'
 require 'action_dispatch'