+++ spec/telegram/bot/rspec/integration/poller_spec.rb
@@ -1,0 +1,5 @@
+require 'telegram/bot/rspec/integration/poller'
+
+RSpec.describe 'Poller integration spec', telegram_bot: :poller do
+  include_examples 'shared integration examples'
+end